feat: renaming UI components following new pattern
