+++ ui/web/src/components/InRoom/elements/Question/Question.scss
